requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
google-api-core==1.25.1
google-auth==1.24.0
google-cloud==0.34.0
google-cloud-core==1.5.0
google-cloud-datastore==2.1.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.16
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.5
pandas==1.2.1
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
